+++ platform-sdk/platform-apps/demos/CryptocurrencyDemo/src/main/java/com/swirlds/demo/crypto/CryptocurrencyDemoMain.java
@@ -39,8 +39,8 @@ import com.swirlds.common.system.SwirldState;
 import com.swirlds.common.threading.framework.StoppableThread;
 import com.swirlds.common.threading.framework.config.StoppableThreadConfiguration;
 import com.swirlds.common.utility.AutoCloseableWrapper;
+import com.swirlds.gui.GuiModel;
 import com.swirlds.platform.Browser;
-import com.swirlds.platform.gui.GuiPlatformAccessor;
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
 import java.util.Random;
@@ -142,10 +142,9 @@ public class CryptocurrencyDemoMain implements SwirldMain {
     public void init(final Platform platform, final NodeId id) {
         this.platform = platform;
         this.selfId = id;
-        final int winNum = GuiPlatformAccessor.getInstance().getInstanceNumber(selfId);
+        final int winNum = GuiModel.getInstance().getInstanceNumber(selfId);
         this.console = createConsole(platform, winNum, true); // create the window, make it visible
-        GuiPlatformAccessor.getInstance()
-                .setAbout(platform.getSelfId(), "Cryptocurrency and stock market demo v. 1.0\n");
+        GuiModel.getInstance().setAbout(platform.getSelfId(), "Cryptocurrency and stock market demo v. 1.0\n");
         this.console.addKeyListener(keyListener);
     }
