+++ platform-sdk/swirlds-common/src/main/java/module-info.java
@@ -151,7 +151,8 @@ module com.swirlds.common {
             com.swirlds.demo.platform,
             com.swirlds.jrs,
             com.swirlds.platform,
-            com.swirlds.platform.test;
+            com.swirlds.platform.test,
+            com.swirlds.gui;
 
     requires transitive com.swirlds.base;
     requires com.swirlds.config;
