+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/Browser.java
@@ -79,6 +79,7 @@ import com.swirlds.config.api.Configuration;
 import com.swirlds.config.api.ConfigurationBuilder;
 import com.swirlds.config.api.source.ConfigSource;
 import com.swirlds.fchashmap.config.FCHashMapConfig;
+import com.swirlds.gui.GuiModel;
 import com.swirlds.gui.WindowConfig;
 import com.swirlds.jasperdb.config.JasperDbConfig;
 import com.swirlds.logging.payloads.NodeAddressMismatchPayload;
@@ -98,7 +99,6 @@ import com.swirlds.platform.event.preconsensus.PreconsensusEventStreamConfig;
 import com.swirlds.platform.event.tipset.EventCreationConfig;
 import com.swirlds.platform.gossip.chatter.config.ChatterConfig;
 import com.swirlds.platform.gossip.sync.config.SyncConfig;
-import com.swirlds.platform.gui.GuiPlatformAccessor;
 import com.swirlds.platform.gui.internal.InfoApp;
 import com.swirlds.platform.gui.internal.InfoMember;
 import com.swirlds.platform.gui.internal.InfoSwirld;
@@ -313,7 +313,7 @@ public class Browser {
                 startPlatforms(configuration, appDefinition, appMains);
 
                 // create the browser window, which uses those Statistics objects
-                showBrowserWindow();
+                showBrowserWindow(null);
                 for (final Frame f : Frame.getFrames()) {
                     if (!f.equals(getBrowserWindow())) {
                         f.toFront();
@@ -693,9 +693,9 @@ public class Browser {
                                 initialState.get(), addressBookInitializer.getInitialAddressBook());
                     }
 
-                    GuiPlatformAccessor.getInstance().setPlatformName(nodeId, platformName);
-                    GuiPlatformAccessor.getInstance().setSwirldId(nodeId, appDefinition.getSwirldId());
-                    GuiPlatformAccessor.getInstance().setInstanceNumber(nodeId, instanceNumber);
+                    GuiModel.getInstance().setPlatformName(nodeId, platformName);
+                    GuiModel.getInstance().setSwirldId(nodeId, appDefinition.getSwirldId());
+                    GuiModel.getInstance().setInstanceNumber(nodeId, instanceNumber);
 
                     platform = new SwirldsPlatform(
                             platformContext,
