+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/WinTab2Consensus.java
@@ -18,8 +18,9 @@ package com.swirlds.platform.gui.internal;
 
 import static com.swirlds.gui.GuiUtils.wrap;
 
+import com.swirlds.gui.GuiModel;
 import com.swirlds.gui.GuiUtils;
-import com.swirlds.gui.PrePaintableJPanel;
+import com.swirlds.gui.components.PrePaintableJPanel;
 import com.swirlds.platform.Consensus;
 import com.swirlds.platform.SwirldsPlatform;
 import com.swirlds.platform.components.state.StateManagementComponent;
@@ -57,7 +58,7 @@ class WinTab2Consensus extends PrePaintableJPanel {
             }
             SwirldsPlatform platform = WinBrowser.memberDisplayed.getPlatform();
             String s = "";
-            s += GuiPlatformAccessor.getInstance().getPlatformName(platform.getSelfId());
+            s += GuiModel.getInstance().getPlatformName(platform.getSelfId());
             final Consensus consensus = GuiPlatformAccessor.getInstance().getConsensus(platform.getSelfId());
             long r1 = consensus.getDeleteRound();
             long r2 = consensus.getFameDecidedBelow();
