+++ platform-sdk/swirlds-platform-gui/src/main/java/com/swirlds/gui/components/ChartLabelModel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020-2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,23 @@
  * limitations under the License.
  */
 
-plugins {
-    id("com.swirlds.platform.conventions")
-    id("com.swirlds.platform.library")
-    id("com.swirlds.platform.maven-publish")
-    id("org.gradle.java-test-fixtures")
-}
+package com.swirlds.gui.components;
 
-dependencies {
-    api(project(":swirlds-base"))
-    implementation(project(":swirlds-common"))
-    implementation(project(":swirlds-logging"))
-    implementation(libs.commons.lang3)
-    compileOnly(libs.spotbugs.annotations)
+public interface ChartLabelModel {
+
+    void setYLabels(String[] labels);
+
+    void setYLabelValues(double[] values);
+
+    void setXLabels(String[] labels);
+
+    void setXLabelValues(double[] values);
+
+    String[] getYLabels();
+
+    double[] getYLabelValues();
+
+    String[] getXLabels();
+
+    double[] getXLabelValues();
 }
